delete exit function at punish
To keep it consistent with the reward I remove the exit from punish in order to keep having the same color (black or white accordingly) on the InterTrial instead of using the unshow that change it to the background color

I removed the exit function from Punish instead of adding an exit on Reward. I think like this is better for the animal because it knows that it cannot start a new trial until the IntreTrial is finished.