change flag initialization
This flag was used for the case where the stim_period is not in current condition to return and not show any stimulus.  When the stim_period='' it also not in the curr_cond but we want to have stimulus show we have to set flag True inside the if statement.

If I remember correct this has been done for a case of SmellyObject stimuli when we wanted to have the odor without the visual stimuli. By not have it on the curr_cond we and adding this flag in Pandas Stimuli we didn't had any visual stimuli.(Smelly object add on the start of pandas the odor)

I think this flag is not correct and we should  remove it. Because it is used in a very specific situation which can be solved on the configuration file in an other way or in an other way not inside the panda Stimuli. ( By the way I had add this flag but now I think it shouldn't be there that i am thinking it again)