Make stim.unshow work at entry of Pretrial
The first time the session starts and go to PreTrial we want an indication for the mouse to understand that the trial has begun so we must use the stim.show.

But the stim.unshow() in Pandas for some reason doesn't work properly the first time so I created the set_taskMgr because taskMgr needs time sleep between steps or to run many steps.