add queue to save responses
Because log_activity is running in the callbacks of interface (multi threading) the get_response can be interrupted from log_activity and after read the last_response overwrite it at the if statement in clean.

So we are using a Queue to save the activity instead in order to be thread safe.

The size of the Queue is small (max_size=3) since the the get_response is called more often than log and we want to create to create a big array and slow down the get_response because it must running fast .