fix on trials_selection =  staircase
iter is initialized at the decleration of ExperimentClass line 33 core/Experiment.py so when a new session begins after sleeping it keeps the last value of iter. So instead I initialize self.iter at the setup function of ExperimentClass.

At line 218 core/Experiment.py when self.iter=1 it doesn't check it the last trial is abort before decide if it must move on the next difficulty or not. At line 219 core/Experiment.py I check if there are some choices and the last one is not abort only then check the performace.