Change end_time at Hydrate state
When end/start time changes and the experiment is at Hydrate State it goes to Offtime. In this case the beh.is_hydrated() returns False and go to Exit and starts a new session. Otherwise it stuck at Offtime because setup_status is running and self.beh.is_sleep_time() returns False